requirements.txt
scikit-image
torch
torchvision
torchaudio

imageio
numpy
matplotlib
tqdm

rospkg

opencv-python
trimesh
open3d
PyMCubes